Extract card scheme image provider
